Create nim diceware cli test program repo
Created an implementation of a Diceware tool in the command line.
Is relatively naive given this is my first nim app, but implements
a good number of command line switches and arguments.

Current implementation will create a passphrase of variable word
length (defaults to 6) from a built-in (or user provided) list of
words.

Dice rolls can be rolled by a built-in pseudo-RNG provided
as command line arguments to the app, or prompted from the user.

More information should probably go in a README...